feature23. New separate file for basic_info application
