+++ src/bot/constants/text.py
@@ -41,11 +41,7 @@ OUR_TEAM_MESSAGE = (
     ' в разделе "Список контактов" или почитать'
     ' об отделах Фонда в разделе "Отделы Фонда"'
 )
-# OUR_TEAM_SECOND_MESSAGE = (
-#     'Также ты можешь посмотреть контакты всех сотрудников'
-#     ' в разделе "Список контактов" или почитать'
-#     ' об отделах Фонда в разделе "Отделы Фонда"'
-# )
+
 ORGANIZATION_MESSAGE = (
     "Как и в любой организации, в Фонде существует"
     ' <a href="https://docs.google.com/spreadsheets/d/1Wr8z0c1QtYvEl1nNX7oOt3LiaQ4ASNBhF2xxDDkpQs8/edit?usp=sharing">организационная структура</a>.'
